% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dm_plotROC_TPRFDR.R
\name{calculate_ROCx}
\alias{calculate_ROCx}
\title{Calculate data for ROC plot}
\usage{
calculate_ROCx(results, status)
}
\arguments{
\item{results}{list of data.frames with "gene_id", "pvalue" and "adj_pvalue" columns}

\item{status}{data.frame with "gene_id" and "status" columns. "status" must consists of 0 and 1 (or FALSE, TRUE)}
}
\description{
Calculate TPR and FPR for ROC plot
}

