% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dm_plotROC_TPRFDR.R
\name{calculate_TPRFDR}
\alias{calculate_TPRFDR}
\title{Calculate data for TPR versus achieved FDR plot}
\usage{
calculate_TPRFDR(results, status, thresholds = c(0.01, 0.05, 0.1))
}
\arguments{
\item{results}{list of data.frames with "gene_id" and "adj_pvalue" columns}

\item{status}{data.frame with "gene_id" and "status" columns. "status" must consists of 0 and 1 (or FALSE, TRUE)}
}
\description{
Calculate TPR and FDR.
}

