% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dm_plotTPRFDR.R
\name{calculateTPRFDR}
\alias{calculateTPRFDR}
\alias{plotTPRFDR}
\title{TPR versus achieved FDR plot}
\usage{
calculateTPRFDR(results, status, thresholds = c(0.01, 0.05, 0.1))

plotTPRFDR(data_TPRFDR, metadata, plot_var, facet_var = NULL,
  plot_colors = NULL, xylim_one = FALSE)
}
\arguments{
\item{results}{List of data.frames with results for different methods. Each data.frame should contain \code{gene_id}, \code{pvalue} and \code{adj_pvalue} columns. Otherwise NAs are returned.}

\item{status}{Data.frame with truth. It must have \code{gene_id} and \code{status} columns, where \code{status} can take values of 0 and 1 (or FALSE, TRUE).}

\item{thresholds}{Thresholds for the FDR.}

\item{data_TPRFDR}{List structured like the output of \code{calculateTPRFDR}.}

\item{metadata}{Data.frame where each row corresponds to one element of \code{data_ROCx} list. Columns describe the origin of this results, for example, which method was used to obtain them.}

\item{plot_var}{Name of one of the variables in \code{metadata} that indicates which results should be displayed in a single panel.}

\item{facet_var}{Vector of one or two variables in \code{metadata} which indicates how the results should be splitted into different panels. If \code{NULL}, no splitting is done.}

\item{plot_colors}{Vector of colors corresponding to the levels of variables indicated by \code{plot_var}. If \code{NULL}, default colors are used.}

\item{xylim_one}{Logical. Whether to set the x and y limits to (0,1).}
}
\value{
calculateTPRFDR returns a list of data.frames with FDR thresholds and the corresponding TPRs and achieved FDRs.
}
\description{
TPR versus achieved FDR plot
}
\examples{
status <- dataDS_status

d <- dataDS_dmDStest
results <- list()
results[[1]] <- results(d)

metadata <- data.frame(method = "DM")

data_TPRFDR <- calculateTPRFDR(results, status,
thresholds = c(0.01, 0.05, 0.1))

plotTPRFDR(data_TPRFDR, metadata, plot_var = "method", facet_var = NULL,
plot_colors = NULL, xylim_one = TRUE)
}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{plotROCx}}, \code{\link{plotVenn}}
}

