% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DM.R
\docType{data}
\name{dataDS_counts}
\alias{dataDS_counts}
\alias{dataDS_dmDSdata}
\alias{dataDS_dmDSdispersion}
\alias{dataDS_dmDStest}
\alias{dataDS_metadata}
\alias{dataDS_status}
\title{Data for differential splicing analysis}
\format{\code{dataDS_counts} contains exonic bin counts. A data.frame with 7 variables:
\itemize{
  \item \code{group_id}: Exonic bin IDs.
  \item \code{sample_1, sample_2, ...}: Quantification of the exonic bins in 6 samples.
  }

\code{dataDS_metadata} describes the samples. A data.frame with 2 variables:
\itemize{
  \item \code{sample_id}: Sample IDs.
  \item \code{group}: Grouping into two conditions.
}

\code{dataDS_status} contains the information about the differential splicing status of genes. A data.frame with 2 variables:
\itemize{
  \item \code{gene_id}: Gene IDs.
  \item \code{status}: 1 for genes with differential splicing and 0 for genes without DS.
}

\code{dataDS_dmDSdata}, \code{dataDS_dmDSdispersion} and \code{dataDS_dmDStest} are the \code{DM} package objects created through the differential splicing analysis pipeline.}
\source{
Soneson C, Matthes KL, Nowicka M, Law CW, Robinson MD. Differential transcript usage from RNA-seq data : isoform pre-filtering improves performance of count-based methods. 2015.

\url{http://imlspenticton.uzh.ch/robinson_lab/splicing_comparison/}
}
\usage{
dataDS_counts

dataDS_metadata

dataDS_status

dataDS_dmDSdata

dataDS_dmDSdispersion

dataDS_dmDStest
}
\description{
A subset (100 randomly selected genes) of exonic bin counts computed with DEXSeq package in a simulation study based on Drosophila genome. In this study data was generated to mimic an RNA-seq assay for 6 samples separated into 2 conditions (3 versus 3 samples). Differential transcript usage was induced by swapping, between the 2 conditions, the expression of two most abundant transcripts for 1000 genes. Our subset consists of 20 genes with differential splicing (status 1) and 80 without (status 0). For more details about this simulations, see the reference in Source paragraph.
}
\keyword{datasets}

