% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DM.R
\docType{data}
\name{dataSQTL_counts}
\alias{dataSQTL_counts}
\alias{dataSQTL_dmSQTLdata}
\alias{dataSQTL_gene_ranges}
\alias{dataSQTL_genotypes}
\title{Data for sQTL analysis}
\format{\code{dataSQTL_counts} contains transcript expected counts. A data.frame with 93 variables:
\itemize{
  \item \code{gene_id}: Gene IDs.
  \item \code{transcript_id}: Transcript IDs.
  \item \code{NA06984, NA06985, ..., NA12890}: Quantification of transcripts for 91 individuals.
  }


 \code{dataSQTL_gene_ranges} with the information about gene location. A \code{\linkS4class{GRanges}} object with gene ranges.


\code{dataSQTL_genotypes} contains genotypes coded as follows: 0 for ref/ref, 1 for ref/not ref, 2 for not ref/not ref, -1 or NA for missing value. A data.frame with 95 variables:
\itemize{
  \item \code{chr}: Chromosome IDs.
  \item \code{start}: Start of SNP location.
  \item \code{end}: End of SNP location.
  \item \code{snp_id}: SNP IDs.
  \item \code{NA06984, NA06985, ..., NA12890}: Genotypes for 91 individuals.
}


\code{dataSQTL_dmSQTLdata} is a \code{DM} package object created through the sQTL analysis pipeline.}
\source{
Lappalainen T, Sammeth M, Friedländer MR, et al. Transcriptome and genome sequencing uncovers functional variation in humans. Nature. 2013;501(7468):506–11.

Genotypes and transcript quantification were downloaded from http://www.ebi.ac.uk/Tools/geuvadis-das/.

Gene annotation Gencode v12 from http://www.gencodegenes.org/releases/12.html.
}
\usage{
dataSQTL_counts

dataSQTL_genotypes

dataSQTL_gene_ranges

dataSQTL_dmSQTLdata
}
\description{
A subset of data from GEUVADIS project where 465 RNA-seq samples from lymphoblastoid cell lines were obtained. 422 of this samples were sequenced within the 1000 Genome Project Phase 1. Here, we make available a subset of bi-allelic SNPs and transcript expected counts for CEPH (CEU) population that corresponds to 50 randomly selected genes from chromosome 19.
}
\keyword{datasets}

