% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DM.R
\docType{package}
\name{DM}
\alias{DM}
\alias{DM-package}
\title{DM: A package to fit Dirichlet-multinomial model}
\description{
The DM package provides a workflow for DS and SQTL analysis
}
\section{DM functions}{

The DM functions ...
}

