% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSdata.R
\docType{class}
\name{dmDSdata-class}
\alias{[,dmDSdata,ANY,ANY-method}
\alias{counts,dmDSdata-method}
\alias{dmDSdata-class}
\alias{length,dmDSdata-method}
\alias{names,dmDSdata-method}
\alias{plot,dmDSdata,ANY-method}
\alias{samples}
\alias{samples,dmDSdata-method}
\title{dmDSdata object}
\usage{
\S4method{counts}{dmDSdata}(object, ...)

samples(x, ...)

\S4method{samples}{dmDSdata}(x)

\S4method{names}{dmDSdata}(x)

\S4method{length}{dmDSdata}(x)

\S4method{[}{dmDSdata,ANY,ANY}(x, i, j)

\S4method{plot}{dmDSdata,ANY}(x, out_dir = NULL)
}
\description{
dmDSdata contains counts of genomic features, such as exons or transcrips, and sample information needed for differential splicing (DS) analysis. It can be created with function \code{\link{dmDSdata}}.
}
\section{Slots}{

\describe{
\item{\code{counts}}{A \code{\linkS4class{MatrixList}} of counts where rows correspond to genomic features, such as exons or transcripts, that are quantified. Columns correspond to samples. 'MatrixList' is partitioned in a way that each submatrix contains counts for a single gene.}

\item{\code{samples}}{A data.frame with information about samples. It must contain columns with unique sample names (\code{sample_id}) and information about grouping into conditions (\code{group}).}
}}

