% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSdispersion.R
\docType{class}
\name{dmDSdispersion-class}
\alias{dmDSdispersion-class}
\title{Object that extends \code{dmDSdata} by adding dipsersion.}
\description{
Object that extends \code{dmDSdata} by adding dipsersion.
}
\section{Slots}{

\describe{
\item{\code{mean_expression}}{Numeric vector of mean gene expression. When empty equals to NULL.}

\item{\code{common_dispersion}}{Numeric or NULL when not estimated.}

\item{\code{tagwise_dispersion}}{Numeric vector of estimated tagwise dispersion.}
}}

