% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSdispersion.R
\docType{methods}
\name{dmDSdispersion}
\alias{dmDSdispersion}
\alias{dmDSdispersion,dmDSdata-method}
\alias{dmDSdispersion,dmDSdispersion-method}
\title{Estimating the dispersion in Dirichlet-multinomial model.}
\usage{
dmDSdispersion(x, ...)

\S4method{dmDSdispersion}{dmDSdata}(x, mean_expression = TRUE,
  common_dispersion = TRUE, tagwise_dispersion = TRUE, disp_adjust = TRUE,
  disp_mode = "grid", disp_interval = c(0, 1e+05), disp_tol = 1e-08,
  disp_init = 100, disp_init_weirMoM = TRUE, disp_grid_length = 21,
  disp_grid_range = c(-10, 10), disp_moderation = "none",
  disp_prior_df = 10, disp_span = 0.3, prop_mode = "constrOptimG",
  prop_tol = 1e-12, verbose = FALSE,
  BPPARAM = BiocParallel::MulticoreParam(workers = 1))

\S4method{dmDSdispersion}{dmDSdispersion}(x, mean_expression = FALSE,
  common_dispersion = FALSE, tagwise_dispersion = TRUE,
  disp_adjust = TRUE, disp_mode = "grid", disp_interval = c(0, 1e+05),
  disp_tol = 1e-08, disp_init = 100, disp_init_weirMoM = TRUE,
  disp_grid_length = 21, disp_grid_range = c(-10, 10),
  disp_moderation = "none", disp_prior_df = 10, disp_span = 0.3,
  prop_mode = "constrOptimG", prop_tol = 1e-12, verbose = FALSE,
  BPPARAM = BiocParallel::MulticoreParam(workers = 1))
}
\arguments{
\item{x}{\code{\link{dmDSdata}} or \code{\link{dmDSdispersion}} object of counts.}

\item{...}{Parameters needed for the dispersion estimation.}

\item{mean_expression}{Logical. Whether to estimate the mean expression of genes.}

\item{common_dispersion}{Logical. Whether to estimate the common dispersion.}

\item{tagwise_dispersion}{Logical. Whether to estimate the tagwise dispersion.}

\item{disp_adjust}{Logical. Whether to use the adjusted or non-adjusted profile likelihood.}

\item{disp_mode}{Optimization method used to maximize the profile likelihood. Possible values are \code{"optimize", "optim", "constrOptim", "grid"}. See Details.}

\item{disp_interval}{The \code{interval} used by \code{\link{optimize}} function that is used to search for the maximum. Valid when using \code{disp_mode = "optimize"}.}

\item{disp_tol}{The desired accuracy when estimating dispersion.}

\item{disp_init}{Initial dispersion.}

\item{disp_init_weirMoM}{Logical. Whether to use the Weir moment estimator as an initial value for dispersion. If \code{TRUE}, the it overwrites the \code{disp_init} value.}

\item{disp_grid_length}{Length of the search grid. Used when \code{disp_mode = "grid"}.}

\item{disp_grid_range}{Vector of ends of grid interval.}

\item{disp_moderation}{Dispersion moderation method. Applies only when \code{disp_mode = "grid"}.}

\item{disp_prior_df}{Prior degrees of freedom used in moderation.}

\item{disp_span}{Value from 0 to 1 defining the percentage of genes used in smoothing sliding window when calculating the trend.}

\item{prop_mode}{Optimization method used to estimate the proportions. Possible values \code{"constrOptim", "constrOptimG"}.}

\item{prop_tol}{The desired accuracy when estimating dispersion.}

\item{verbose}{Logical.}

\item{BPPARAM}{Parallelization method used by \code{\link[BiocParallel]{bplapply}}.}
}
\value{
Returns the \code{\linkS4class{dmDSdispersion}} object.
}
\description{
Function that is a wrapper for different optimazation methods used to estimate the dispersion parameters of Dirichlet-multinomial model with maximum likelihood. Parameters that are directly used in the dispersion estimation start with prefix \code{disp_} and the one that are used directly for the proportion estimation start with \code{prop_}.
}
\details{
TODO: describe in detail all the methods...

disp_tol: The accuracy defined as \code{tol} in \code{\link{optimize}} when \code{disp_mode = "optimize"}, as \code{factr} in \code{\link{optim}} when \code{disp_mode = "optim"}, as \code{reltol} in \code{\link{constrOptim}} when \code{disp_mode = "constrOptim"}.
}
\examples{
data <- dataDS_dmDSdata
data <- dmDSfilter(data)
\\donotrun{
data <- dmDSdispersion(data)
}
}
\seealso{
\code{\link[BiocParallel]{bplapply}}
}

