% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSdata.R
\docType{methods}
\name{dmDSfilter}
\alias{dmDSfilter}
\alias{dmDSfilter,dmDSdata-method}
\title{Filtering.}
\usage{
dmDSfilter(x, ...)

\S4method{dmDSfilter}{dmDSdata}(x, min_samps_gene_expr = 3,
  min_gene_expr = 1, min_samps_feature_prop = 3, min_feature_prop = 0.01,
  max_features = Inf)
}
\arguments{
\item{x}{\code{\link{dmDSdata}} object of counts.}

\item{...}{Filtering parameters.}

\item{min_samps_gene_expr}{Minimal number of samples where the genes should
be expressed.}

\item{min_gene_expr}{Minimal gene expression in CPM \code{\link[edgeR]{cpm}}.}

\item{min_samps_feature_prop}{Minimal number of samples where the features
should be expressed.}

\item{min_feature_prop}{Minimal proportion for feature expression. This value
should be between 0 and 1.}

\item{max_features}{Maximum number of features that should be kept per gene.}
}
\value{
Returns filtered \code{\linkS4class{dmDSdata}} object.
}
\description{
Filtering of genes with low expression and features with low proportions.
}
\details{
Parameters should be adjusted based on the sample size used for the analysis...
}
\examples{
data <- dataDS_dmDSdata
dmDSplotData(data)
data <- dmDSfilter(data)
dmDSplotData(data)
}

