% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSfit.R
\docType{methods}
\name{dmDSfit}
\alias{dmDSfit}
\alias{dmDSfit,dmDSdispersion-method}
\title{Estimating the proportions and likelihoods of Dirichlet-multinomial full and null models.}
\usage{
dmDSfit(x, ...)

\S4method{dmDSfit}{dmDSdispersion}(x, dispersion = "tagwise_dispersion",
  prop_mode = c("constrOptim", "constrOptimG", "FisherScoring")[2],
  prop_tol = 1e-12, verbose = FALSE,
  BPPARAM = BiocParallel::MulticoreParam(workers = 1))
}
\arguments{
\item{x}{\code{\link{dmDSdispersion}} object or any that inherits from it i.e. \code{\link{dmDSfit}} or \code{\link{dmDStest}}.}

\item{...}{Parameters needed for the proportion estimation.}

\item{dispersion}{Characted defining which dispersion should be used for fitting. Possible values \code{"tagwise_dispersion", "common_dispersion"}}

\item{prop_mode}{Optimization method used to estimate the proportions. Possible values \code{"constrOptim", "constrOptimG"}.}

\item{prop_tol}{The desired accuracy when estimating dispersion.}

\item{verbose}{Logical.}

\item{BPPARAM}{Parallelization method used by \code{\link[BiocParallel]{bplapply}}.}
}
\description{
Estimating the proportions and likelihoods of Dirichlet-multinomial full and null models.
}

