% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSLRT.R
\docType{class}
\name{dmDSLRT-class}
\alias{dmDSLRT-class}
\alias{plot,dmDSLRT,ANY-method}
\alias{results}
\alias{results,dmDSLRT-method}
\title{Object that extends \code{dmDSfit} by adding the test results.}
\usage{
results(x, ...)

\S4method{results}{dmDSLRT}(x)

\S4method{plot}{dmDSLRT,ANY}(x, out_dir = NULL)
}
\description{
Object that extends \code{dmDSfit} by adding the test results.
}
\section{Slots}{

\describe{
\item{\code{fit_null}}{list of null models defined by contrasts.}

\item{\code{results}}{data.frame with \code{gene_id} - gene IDs, \code{lr} - likelihood ratio statistics, \code{df} - degrees of freedom, \code{pvalue} - p-values and \code{adj_pvalue} - Benjamini & Hochberg adjusted p-values.}
}}

