% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSdata.R
\docType{methods}
\name{dmDSplotData}
\alias{dmDSplotData}
\alias{dmDSplotData,dmDSdata-method}
\title{Plot the data information.}
\usage{
dmDSplotData(x, ...)

\S4method{dmDSplotData}{dmDSdata}(x, out_dir = NULL, info = NULL)
}
\arguments{
\item{x}{\code{\link{dmDSdata}} object of counts.}

\item{...}{Plotting parameters.}

\item{out_dir}{Directory where the plot should be saved. If \code{NULL} the plot is printed.}

\item{info}{\code{DataFrame} with \code{gene_id} and \code{feature_id} that are differentially spliced (DS).}
}
\description{
Plots a histogram of number of features per gene.
}
\examples{
dmDSplotData(dmDSdata)
}

