% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSdispersion.R
\docType{methods}
\name{dmDSplotDispersion}
\alias{dmDSplotDispersion}
\alias{dmDSplotDispersion,dmDSdispersion-method}
\title{Plot the dispersion - mean trend.}
\usage{
dmDSplotDispersion(x, ...)

\S4method{dmDSplotDispersion}{dmDSdispersion}(x, out_dir = NULL)
}
\arguments{
\item{x}{\code{\link{dmDSdispersion}} object or any that inherits from it i.e. \code{\link{dmDSfit}} or \code{\link{dmDStest}}.}

\item{...}{Plotting parameters.}

\item{out_dir}{Directory where the plot should be saved. If \code{NULL} the plot is printed.}
}
\description{
Plot the dispersion - mean trend.
}
\examples{
dmDSplotDispersion(dataDS_dmDSdispersion)
}

