% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSfit.R, R/class_dmDStest.R
\docType{methods}
\name{dmDSplotFit}
\alias{dmDSplotFit}
\alias{dmDSplotFit,dmDSfit-method}
\alias{dmDSplotFit,dmDStest-method}
\title{Plot the estimated proportions.}
\usage{
dmDSplotFit(x, ...)

\S4method{dmDSplotFit}{dmDSfit}(x, gene_id, plot_type = "barplot",
  order = TRUE, plot_full = TRUE, plot_null = TRUE, out_dir = NULL)

\S4method{dmDSplotFit}{dmDStest}(x, gene_id, plot_type = c("barplot",
  "boxplot1", "boxplot2", "lineplot", "ribbonplot")[1], order = TRUE,
  plot_full = TRUE, plot_nunll = TRUE, out_dir = NULL)
}
\arguments{
\item{x}{\code{\link{dmDSfit}} object or any that inherits from it i.e. \code{\link{dmDStest}}.}

\item{...}{Plotting parameters.}

\item{gene_id}{Vector of gene IDs to be plotted.}

\item{plot_type}{Character defining type of the plot produced. Possible values \code{"barplot", "boxplot1", "boxplot2", "lineplot", "ribbonplot"}.}

\item{order}{Logical. Whether to plot the features ordered by their expression.}

\item{plot_full}{Logical. Whether to plot the proportions estimated by the full model.}

\item{plot_null}{Logical. Whether to plot the proportions estimated by the null model.}

\item{out_dir}{Directory where the plot should be saved. If \code{NULL} the plot is printed.}
}
\description{
Plot the estimated proportions.
}

