% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDStest.R
\docType{methods}
\name{dmDSplotTest}
\alias{dmDSplotTest}
\alias{dmDSplotTest,dmDStest-method}
\title{Plot the histogram of p-values.}
\usage{
dmDSplotTest(x, ...)

\S4method{dmDSplotTest}{dmDStest}(x, out_dir = NULL)
}
\arguments{
\item{x}{\code{\link{dmDStest}} object.}

\item{...}{Plotting parameters.}

\item{out_dir}{Directory where the plot should be saved. If \code{NULL} the plot is printed.}
}
\description{
Plot the histogram of p-values.
}

