% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDStest.R
\docType{methods}
\name{dmDStest}
\alias{dmDStest}
\alias{dmDStest,dmDSfit-method}
\title{Likelihood ratio test between full and null model.}
\usage{
dmDStest(x, ...)

\S4method{dmDStest}{dmDSfit}(x)
}
\arguments{
\item{x}{\code{\link{dmDStest}} object.}

\item{...}{Parameters needed for the likelihood ratio test.}
}
\value{
This function returns a \code{\link{dmDStest}} object with sorted by significance \code{table} that contains  \code{gene_id} - gene IDs, \code{lr} - likelihood ratio statistics, \code{df} - degrees of freedom, \code{pvalue} - p-values and \code{adj_pvalue} - Benjamini & Hochberg adjusted p-values.
}
\description{
Likelihood ratio test between full and null model.
}

