% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSdata.R, R/class_dmSQTLdata.R
\docType{methods}
\name{dmFilter}
\alias{dmFilter}
\alias{dmFilter,dmDSdata-method}
\alias{dmFilter,dmSQTLdata-method}
\title{Filtering.}
\usage{
dmFilter(x, ...)

\S4method{dmFilter}{dmDSdata}(x, min_samps_gene_expr = 3, min_gene_expr = 1,
  min_samps_feature_prop = 3, min_feature_prop = 0.01, max_features = Inf)

\S4method{dmFilter}{dmSQTLdata}(x, min_samps_gene_expr = 70,
  min_gene_expr = 1, min_samps_feature_prop = 5, min_feature_prop = 0.1,
  max_features = Inf, minor_allel_freq = 0.05,
  BPPARAM = BiocParallel::MulticoreParam(workers = 1))
}
\arguments{
\item{x}{\code{\linkS4class{dmDSdata}} or \code{\linkS4class{dmSQTLdata}} object.}

\item{...}{Filtering parameters.}

\item{min_samps_gene_expr}{Minimal number of samples where the genes should
be expressed.}

\item{min_gene_expr}{Minimal gene expression in CPM \code{\link[edgeR]{cpm}}.}

\item{min_samps_feature_prop}{Minimal number of samples where the features
should be expressed.}

\item{min_feature_prop}{Minimal proportion for feature expression. This value
should be between 0 and 1.}

\item{max_features}{Maximum number of features that should be kept per gene.}

\item{minor_allel_freq}{Value between 0 and 1 which corresponds to minimal
allel frequency.}

\item{BPPARAM}{Parallelization method used by
\code{\link[BiocParallel]{bplapply}}.}
}
\value{
Returns filtered \code{\linkS4class{dmDSdata}} or \code{\linkS4class{dmSQTLdata}} object.
}
\description{
Filtering of genes with low expression and features with low proportions.
}
\details{
Parameters should be adjusted based on the sample size used for the analysis...
}

