% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmFit.R
\docType{class}
\name{dmFit-class}
\alias{dmFit-class}
\title{Object that contains Dirichlet-multinomial fitting results.}
\description{
Object that contains Dirichlet-multinomial fitting results.
}
\section{Slots}{

\describe{
\item{\code{proportions}}{\code{\linkS4class{MatrixList}} of fitted proportions. Length of this list is equal to the number of genes. Each element is a matrix with rows corresponding to the features of a gene and columns to the conditions/grouping of samples.}

\item{\code{statistics}}{DataFrame with per gene loglikelihoods \code{lik} and degrees of freedom \code{df}}
}}

