% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSLRT.R, R/class_dmSQTLLRT.R
\docType{methods}
\name{dmLRT}
\alias{dmLRT}
\alias{dmLRT,dmDSfit-method}
\alias{dmLRT,dmSQTLfit-method}
\title{Likelihood ratio test between full and null model.}
\usage{
dmLRT(x, ...)

\S4method{dmLRT}{dmDSfit}(x)

\S4method{dmLRT}{dmSQTLfit}(x, BPPARAM = BiocParallel::MulticoreParam(workers
  = 1))
}
\arguments{
\item{x}{\code{\linkS4class{dmDSfit}} or \code{\linkS4class{dmSQTLfit}} object.}

\item{...}{Parameters needed for the likelihood ratio test.}
}
\value{
This function returns a \code{\linkS4class{dmDSLRT}} or \code{\linkS4class{dmSQTLLRT}} object with an additional slot \code{table} which is sorted by significance and contains  \code{gene_id} - gene IDs, \code{lr} - likelihood ratio statistics, \code{df} - degrees of freedom, \code{pvalue} - p-values and \code{adj_pvalue} - Benjamini & Hochberg adjusted p-values.
}
\description{
Likelihood ratio test between full and null model.
}

