% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmSQTLdata.R
\docType{class}
\name{dmSQTLdata-class}
\alias{[,dmSQTLdata,ANY,ANY-method}
\alias{dmSQTLdata-class}
\alias{length,dmSQTLdata-method}
\alias{names,dmSQTLdata-method}
\alias{plot,dmSQTLdata,ANY-method}
\title{Object that contains counts, genotypes and sample information.}
\usage{
\S4method{names}{dmSQTLdata}(x)

\S4method{length}{dmSQTLdata}(x)

\S4method{[}{dmSQTLdata,ANY,ANY}(x, i, j)

\S4method{plot}{dmSQTLdata,ANY}(x, out_dir = NULL)
}
\description{
Can be created with function \code{\link{dmSQTLdata}} and \code{\link{dmSQTLdataFromRanges}}.
}
\section{Slots}{

\describe{
\item{\code{counts}}{\code{\linkS4class{MatrixList}}  of counts.}

\item{\code{genotypes}}{\code{\linkS4class{MatrixList}} with unique genotypes.}

\item{\code{blocks}}{\code{\linkS4class{MatrixList}} with blocks of genotypes.}

\item{\code{samples}}{data.frame with information about samples. Contains unique sample names (\code{sample_id}).}
}}

