% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmSQTLdata.R
\name{dmSQTLdata}
\alias{dmSQTLdata}
\title{Create \code{\linkS4class{dmSQTLdata}} object from tables of counts and matched genotypes}
\usage{
dmSQTLdata(counts, gene_id_counts, feature_id_counts, genotypes,
  gene_id_genotypes, snp_id_genotypes, sample_id)
}
\arguments{
\item{counts}{A numeric matrix of counts. Rows represent features (exons,
  bins or transcripts), columns represent samples.}

\item{gene_id_counts}{Vector of gene IDs of lenght correspoding to the number of rows in \code{counts}.}

\item{feature_id_counts}{Vector of feature IDs of lenght correspoding to the number of rows in \code{counts}.}

\item{genotypes}{A numeric matrix of matched genotypes. See Details.}

\item{gene_id_genotypes}{Vector of gene IDs that are matched with genotypes.}

\item{snp_id_genotypes}{Vector of SNP IDs that correspond to genotypes.}

\item{sample_id}{A vector of unique sample IDs of length corresponding to
  the number of columns in \code{counts}.}
}
\value{
Returns a \code{\linkS4class{dmSQTLdata}} object containing counts, genotypes and sample
   information.
}
\description{
Create \code{\linkS4class{dmSQTLdata}} object from tables of counts and matched genotypes
}

