% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmSQTLfit.R
\docType{class}
\name{dmSQTLfit-class}
\alias{dmSQTLfit-class}
\alias{plot,dmSQTLfit,ANY-method}
\title{Object that extends \code{dmSQTLdispersion} by adding fitting.}
\usage{
\S4method{plot}{dmSQTLfit,ANY}(x, gene_id, snp_id, plot_type = "boxplot1",
  order = TRUE, plot_full = TRUE, plot_main = TRUE, out_dir = NULL)
}
\description{
Object that extends \code{dmSQTLdispersion} by adding fitting.
}
\section{Slots}{

\describe{
\item{\code{dispersion}}{Character specifying which type of dispersion was used for fitting.}

\item{\code{fit_full}}{List of \code{\linkS4class{MatrixList}} objects. Each of them contains the per gene fitting results that include the group proportions (groups are defined by genotypes) of feature expression, full model likelihoods and degrees of freedom.}
}}

