% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmSQTLfit.R
\docType{methods}
\name{dmSQTLfit}
\alias{dmSQTLfit}
\alias{dmSQTLfit,dmSQTLdispersion-method}
\title{Estimating the proportions and likelihoods of Dirichlet-multinomial full and null models.}
\usage{
dmSQTLfit(x, ...)

\S4method{dmSQTLfit}{dmSQTLdispersion}(x, dispersion = "tagwise_dispersion",
  prop_mode = c("constrOptim", "constrOptimG", "FisherScoring")[2],
  prop_tol = 1e-12, verbose = FALSE,
  BPPARAM = BiocParallel::MulticoreParam(workers = 1))
}
\arguments{
\item{x}{\code{\link{dmSQTLdispersion}} object or any that inherits from it i.e. \code{\link{dmSQTLfit}} or \code{\link{dmSQTLtest}}.}

\item{...}{Parameters needed for the proportion estimation.}

\item{dispersion}{Characted defining which dispersion should be used for fitting. Possible values \code{"tagwise_dispersion", "common_dispersion"}}

\item{prop_mode}{Optimization method used to estimate the proportions. Possible values \code{"constrOptim", "constrOptimG"}.}

\item{prop_tol}{The desired accuracy when estimating dispersion.}

\item{verbose}{Logical.}

\item{BPPARAM}{Parallelization method used by \code{\link[BiocParallel]{bplapply}}.}
}
\description{
Estimating the proportions and likelihoods of Dirichlet-multinomial full and null models.
}
\examples{
data <- dataSQTL_dmSQTLdispersion
data <- dmSQTLfit(data)

snp_id <- "snp_19_12796435"
gene_id <- "ENSG00000132004.7"

dmSQTLplotFit(data, gene_id, snp_id)
}

