% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmSQTLLRT.R
\docType{class}
\name{dmSQTLLRT-class}
\alias{dmSQTLLRT-class}
\alias{plot,dmSQTLLRT,ANY-method}
\alias{results,dmSQTLLRT-method}
\title{Object that extends \code{dmSQTLfit} by adding the test results.}
\usage{
\S4method{results}{dmSQTLLRT}(x)

\S4method{plot}{dmSQTLLRT,ANY}(x, out_dir = NULL)
}
\description{
Object that extends \code{dmSQTLfit} by adding the test results.
}
\section{Slots}{

\describe{
\item{\code{fit_null}}{list}

\item{\code{results}}{data.frame with \code{gene_id} - gene IDs, \code{snp_id} - SNP IDs, \code{lr} - likelihood ratio statistics, \code{df} - degrees of freedom, \code{pvalue} - p-values and \code{adj_pvalue} - Benjamini & Hochberg adjusted p-values.}
}}

