% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmSQTLdata.R
\docType{methods}
\name{dmSQTLplotData}
\alias{dmSQTLplotData}
\alias{dmSQTLplotData,dmSQTLdata-method}
\title{Plot the data information.}
\usage{
dmSQTLplotData(x, ...)

\S4method{dmSQTLplotData}{dmSQTLdata}(x, out_dir = NULL)
}
\arguments{
\item{x}{\code{\link{dmSQTLdata}} object of counts.}

\item{...}{Plotting parameters.}

\item{out_dir}{Directory where the plot should be saved. If \code{NULL} the plot is printed.}
}
\description{
Plots histograms of number of features per gene and number of SNPs per gene.
}

