% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmSQTLdispersion.R
\docType{methods}
\name{dmSQTLplotDispersion}
\alias{dmSQTLplotDispersion}
\alias{dmSQTLplotDispersion,dmSQTLdispersion-method}
\title{Plot the dispersion - mean trend.}
\usage{
dmSQTLplotDispersion(x, ...)

\S4method{dmSQTLplotDispersion}{dmSQTLdispersion}(x, out_dir = NULL)
}
\arguments{
\item{x}{\code{\link{dmSQTLdispersion}} object or any that inherits from it i.e. \code{\link{dmSQTLfit}} or \code{\link{dmSQTLtest}}.}

\item{...}{Plotting parameters.}

\item{out_dir}{Directory where the plot should be saved. If \code{NULL} the plot is printed.}
}
\description{
Plot the dispersion - mean trend.
}

