% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmSQTLfit.R, R/class_dmSQTLtest.R
\docType{methods}
\name{dmSQTLplotFit}
\alias{dmSQTLplotFit}
\alias{dmSQTLplotFit,dmSQTLfit-method}
\alias{dmSQTLplotFit,dmSQTLtest-method}
\title{Plot the estimated proportions.}
\usage{
dmSQTLplotFit(x, ...)

\S4method{dmSQTLplotFit}{dmSQTLfit}(x, gene_id, snp_id,
  plot_type = "boxplot1", order = TRUE, plot_full = TRUE,
  plot_null = TRUE, out_dir = NULL)

\S4method{dmSQTLplotFit}{dmSQTLtest}(x, gene_id, snp_id,
  plot_type = "boxplot1", order = TRUE, plot_full = TRUE,
  plot_null = TRUE, out_dir = NULL)
}
\arguments{
\item{x}{\code{\link{dmSQTLfit}} object or any that inherits from it i.e. \code{\link{dmSQTLtest}}.}

\item{...}{Plotting parameters.}

\item{gene_id}{Vector of gene IDs to be plotted.}

\item{snp_id}{Vector of SNP IDs to be plotted.}

\item{plot_type}{Character defining type of the plot produced. Possible values \code{"barplot", "boxplot1", "boxplot2", "lineplot", "ribbonplot"}.}

\item{order}{Logical. Whether to plot the features ordered by their expression.}

\item{plot_full}{Logical. Whether to plot the proportions estimated by the full model.}

\item{plot_null}{Logical. Whether to plot the proportions estimated by the null model.}

\item{out_dir}{Directory where the plot should be saved. If \code{NULL} the plot is printed.}
}
\description{
Plot the estimated proportions.
}

