% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmSQTLtest.R
\docType{methods}
\name{dmSQTLplotTest}
\alias{dmSQTLplotTest}
\alias{dmSQTLplotTest,dmSQTLtest-method}
\title{Plot the histogram of p-values.}
\usage{
dmSQTLplotTest(x, ...)

\S4method{dmSQTLplotTest}{dmSQTLtest}(x, out_dir = NULL)
}
\arguments{
\item{x}{\code{\link{dmSQTLtest}} object.}

\item{...}{Plotting parameters.}

\item{out_dir}{Directory where the plot should be saved. If \code{NULL} the plot is printed.}
}
\description{
Plot the histogram of p-values.
}

