% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmSQTLtest.R
\docType{methods}
\name{dmSQTLtest}
\alias{dmSQTLtest}
\alias{dmSQTLtest,dmSQTLfit-method}
\title{Likelihood ratio test between full and null model.}
\usage{
dmSQTLtest(x, ...)

\S4method{dmSQTLtest}{dmSQTLfit}(x,
  BPPARAM = BiocParallel::MulticoreParam(workers = 1))
}
\arguments{
\item{x}{\code{\link{dmSQTLtest}} object.}

\item{...}{Parameters needed for the likelihood ratio test.}

\item{BPPARAM}{Parallelization method used by \code{\link[BiocParallel]{bplapply}}.}
}
\value{
This function returns a \code{\link{dmSQTLtest}} object with an additional slot \code{table} which is sorted by significance and contains  \code{gene_id} - gene IDs, \code{snp_id} - SNP IDs, \code{lr} - likelihood ratio statistics, \code{df} - degrees of freedom, \code{pvalue} - p-values and \code{adj_pvalue} - Benjamini & Hochberg adjusted p-values.
}
\description{
Likelihood ratio test between full and null model.
}
\examples{
data <- dataSQTL_dmSQTLdispersion
data <- dmSQTLfit(data)

data <- dmSQTLtest(data)
dmSQTLplotTest(data)

snp_id <- "snp_19_48981946"
gene_id <- "ENSG00000105443.8"
dmSQTLplotFit(data, gene_id, snp_id)
}

