% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDStest.R, R/class_dmSQTLtest.R
\docType{methods}
\name{dmTest}
\alias{dmTest}
\alias{dmTest,dmDSfit-method}
\alias{dmTest,dmSQTLfit-method}
\title{Likelihood ratio test between full and null model.}
\usage{
dmTest(x, ...)

\S4method{dmTest}{dmDSfit}(x, compared_groups = 1:nlevels(samples(x)$group),
  prop_mode = "constrOptimG", prop_tol = 1e-12, verbose = FALSE,
  BPPARAM = BiocParallel::MulticoreParam(workers = 1))

\S4method{dmTest}{dmSQTLfit}(x, prop_mode = "constrOptimG",
  prop_tol = 1e-12, verbose = FALSE,
  BPPARAM = BiocParallel::MulticoreParam(workers = 1))
}
\arguments{
\item{x}{\code{\linkS4class{dmDSfit}} or \code{\linkS4class{dmSQTLfit}} object.}

\item{...}{Parameters needed for the likelihood ratio test.}

\item{compared_groups}{vector or a list of vectors.}

\item{prop_mode}{Optimization method used to estimate proportions. Possible values \code{"constrOptim", "constrOptimG"}.}

\item{prop_tol}{The desired accuracy when estimating proportions.}

\item{verbose}{Logical. Wheter to display more progress messages.}

\item{BPPARAM}{Parallelization method used by \code{\link[BiocParallel]{bplapply}}.}
}
\value{
This function returns a \code{\linkS4class{dmDStest}} or \code{\linkS4class{dmSQTLtest}} object with an additional slot \code{table} which is sorted by significance and contains  \code{gene_id} - gene IDs, \code{lr} - likelihood ratio statistics, \code{df} - degrees of freedom, \code{pvalue} - p-values and \code{adj_pvalue} - Benjamini & Hochberg adjusted p-values.
}
\description{
Likelihood ratio test between full and null model.
}
\examples{
d <- dataDS_dmDSdispersion

# If possible, increase the number of workers
d <- dmFit(d, BPPARAM = BiocParallel::MulticoreParam(workers = 1))

d <- dmTest(d)

results <- results(d)
}

