% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_MatrixList.R
\docType{class}
\name{MatrixList-class}
\alias{MatrixList-class}
\title{MatrixList container}
\description{
MatrixList container
}
\section{Slots}{

\describe{
\item{\code{unlistData}}{Matrix data}

\item{\code{partitioning}}{List with indexes of each submatrix}

\item{\code{metadata}}{Matrix of additional information}
}}

