% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_MatrixList.R
\name{MatrixList}
\alias{MatrixList}
\title{Create a MatrixList object from a list of matrices}
\usage{
MatrixList(...)
}
\arguments{
\item{...}{A list of numeric matrices with the same number of columns. NULL is allowed.}
}
\description{
Create a MatrixList object from a list of matrices
}

