% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSdata.R, R/class_dmSQTLdata.R
\docType{methods}
\name{plotData}
\alias{plotData}
\alias{plotData,dmDSdata-method}
\alias{plotData,dmSQTLdata-method}
\title{Plot the data information.}
\usage{
plotData(x, ...)

\S4method{plotData}{dmDSdata}(x, out_dir = NULL)

\S4method{plotData}{dmSQTLdata}(x, out_dir = NULL)
}
\arguments{
\item{x}{\code{\linkS4class{dmDSdata}} or \code{\linkS4class{dmSQTLdata}} object with data.}

\item{...}{Plotting parameters.}

\item{out_dir}{Directory where the plot should be saved. If \code{NULL} the plot is printed.}
}
\description{
Plots a histogram of number of features per gene.
}
\examples{
### dmDSdata

d <- dataDS_dmDSdata
plotData(d)
plot(d)

### dmSQTLdata

d <- dataSQTL_dmSQTLdata
plotData(d)
plot(d)
}

