% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSdispersion.R, R/class_dmSQTLdispersion.R
\docType{methods}
\name{plotDispersion}
\alias{plotDispersion}
\alias{plotDispersion,dmDSdispersion-method}
\alias{plotDispersion,dmSQTLdispersion-method}
\title{Dispersion versus mean expression plot}
\usage{
plotDispersion(x, ...)

\S4method{plotDispersion}{dmDSdispersion}(x, out_dir = NULL)

\S4method{plotDispersion}{dmSQTLdispersion}(x, out_dir = NULL)
}
\arguments{
\item{x}{\code{\linkS4class{dmDSdispersion}} or \code{\linkS4class{dmSQTLdispersion}} object.}

\item{...}{Other parameters that can be defined by methods using this generic.}

\item{out_dir}{Directory where the plot should be saved. If \code{NULL} the plot is printed.}
}
\description{
Dispersion versus mean expression plot
}
\examples{
d <- dataDS_dmDSdispersion
plotDispersion(d)
plot(d)
}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{plotData}}, \code{\link{plotFit}}, \code{\link{plotLRT}}
}

