% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSLRT.R, R/class_dmDSfit.R, R/class_dmSQTLLRT.R, R/class_dmSQTLfit.R
\docType{methods}
\name{plotFit,dmDSLRT-method}
\alias{plotFit}
\alias{plotFit,dmDSLRT-method}
\alias{plotFit,dmDSfit-method}
\alias{plotFit,dmSQTLLRT-method}
\alias{plotFit,dmSQTLfit-method}
\title{Plot the estimated proportions.}
\usage{
\S4method{plotFit}{dmDSLRT}(x, compared_groups = 1, gene_id,
  plot_type = c("barplot", "boxplot1", "boxplot2", "lineplot",
  "ribbonplot")[1], order = TRUE, plot_full = TRUE, plot_null = TRUE,
  plot_main = TRUE, out_dir = NULL)

plotFit(x, ...)

\S4method{plotFit}{dmDSfit}(x, gene_id, plot_type = "barplot", order = TRUE,
  plot_full = TRUE, plot_main = TRUE, out_dir = NULL)

\S4method{plotFit}{dmSQTLLRT}(x, gene_id, snp_id, plot_type = "boxplot1",
  order = TRUE, plot_full = TRUE, plot_null = TRUE, plot_main = TRUE,
  out_dir = NULL)

\S4method{plotFit}{dmSQTLfit}(x, gene_id, snp_id, plot_type = "boxplot1",
  order = TRUE, plot_full = TRUE, plot_main = TRUE, out_dir = NULL)
}
\arguments{
\item{x}{\code{\linkS4class{dmDSfit}} or \code{\linkS4class{dmSQTLfit}} object or any that inherits from them.}

\item{compared_groups}{numeric or character}

\item{gene_id}{Vector of gene IDs to be plotted.}

\item{plot_type}{Character defining type of the plot produced. Possible values \code{"barplot"}, \code{"boxplot1"}, \code{"boxplot2"}, \code{"lineplot"}, \code{"ribbonplot"}.}

\item{order}{Logical. Whether to plot the features ordered by their expression.}

\item{plot_full}{Logical. Whether to plot the proportions estimated by the full model.}

\item{plot_main}{Logical. Whether to plot the plot title.}

\item{...}{Plotting parameters.}
}
\description{
Plot the estimated proportions.
}
\examples{
d <- dataDS_dmDSdispersion

# If possible, increase the number of workers
d <- dmFit(d, BPPARAM = BiocParallel::MulticoreParam(workers = 1))

gene_id <- names(d)[1]

plotFit(d, gene_id = gene_id)
plot(d, gene_id = gene_id, plot_type = "lineplot", plot_full = FALSE)

d <- dmLRT(d)

results <- results(d)

gene_id <- results$gene_id[1:3]

plotFit(d, gene_id = gene_id)
}

