% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSLRT.R, R/class_dmSQTLLRT.R
\docType{methods}
\name{plotLRT}
\alias{plotLRT}
\alias{plotLRT,dmDSLRT-method}
\alias{plotLRT,dmSQTLLRT-method}
\title{Plot the histogram of p-values.}
\usage{
plotLRT(x, ...)

\S4method{plotLRT}{dmDSLRT}(x, out_dir = NULL)

\S4method{plotLRT}{dmSQTLLRT}(x, out_dir = NULL)
}
\arguments{
\item{x}{\code{\linkS4class{dmDSLRT}} or \code{\linkS4class{dmSQTLLRT}} object.}

\item{...}{Plotting parameters.}
}
\description{
Plot the histogram of p-values.
}

