% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dm_plotVenn.R
\name{plotVenn}
\alias{calculateVenn}
\alias{plotVenn}
\title{Venn diagram}
\usage{
calculateVenn(results, status = NULL, threshold = 0.05)

plotVenn(data_venn, plot_results, metadata, plot_var, plot_colors = NULL,
  plot_status = TRUE)
}
\arguments{
\item{results}{List of data frames with results from different methods. Each data frame should contain \code{gene_id}, \code{pvalue} and \code{adj_pvalue} columns. Otherwise NAs are returned.}

\item{status}{Data frame with truth. It must have \code{gene_id} and \code{status} columns, where \code{status} can take values of 0 and 1 (or FALSE, TRUE).}

\item{threshold}{Threshold for the FDR.}

\item{data_venn}{List structured like the output of \code{calculateVenn}.}

\item{plot_results}{Numeric vector of maximum length 4 (if \code{plot_status = TRUE}) or 5 (if \code{plot_status = FALSE}) indicating which results should be plotted.}

\item{metadata}{Data frame where each row corresponds to one element of \code{data_ROCx} list. Columns describe the origin of this results, for example, which method was used to obtain them.}

\item{plot_var}{Name of one of the variables in \code{metadata}. Levels of this variable will be used as category names in venn diagram.}

\item{plot_colors}{Vector of colors corresponding to the levels of variables indicated by \code{plot_var}. If \code{NULL}, default colors are used.}

\item{plot_status}{Logical. Whether to plot a circle that corresponds to truly significant genes.}
}
\value{
\code{calculateVenn} returns a list of vectors with names of the genes that are significant \code{adj_pvalue < threshold} for different methods. If status is not \code{NULL}, the last element of the output is a list of genes with status equal 1.

\code{plotVenn} returns a venn diagram overlaps of significant genes for different methods. \code{plotVenn} is a wrapper function that uses \code{\link[VennDiagram]{venn.diagram}} for plotting.
}
\description{
Venn diagram overlaps for a list of result data frames.
}
\examples{
status <- dataDS_status
d <- dataDS_dmDStest

results <- list()
results[[1]] <- results(d)
metadata <- data.frame(method = "DM")

data_venn <- calculateVenn(results, status = status, threshold = 0.05)

plotVenn(data_venn, plot_results = 1, metadata, plot_var = "method",
   plot_colors = NULL, plot_status = TRUE)
}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{dataDS_status}}, \code{\link{dataDS_dmDStest}}, \code{\link{plotROCx}}, \code{\link{plotTPRFDR}}
}

