% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSdata.R
\docType{class}
\name{dmDSdata-class}
\alias{[,dmDSdata,ANY,ANY-method}
\alias{counts}
\alias{counts,dmDSdata-method}
\alias{dmDSdata-class}
\alias{length,dmDSdata-method}
\alias{names,dmDSdata-method}
\alias{plot,dmDSdata,ANY-method}
\alias{samples}
\alias{samples,dmDSdata-method}
\title{Object that contains counts and sample information.}
\usage{
counts(x, ...)

\S4method{counts}{dmDSdata}(x)

samples(x, ...)

\S4method{samples}{dmDSdata}(x)

\S4method{names}{dmDSdata}(x)

\S4method{length}{dmDSdata}(x)

\S4method{[}{dmDSdata,ANY,ANY}(x, i, j)

\S4method{plot}{dmDSdata,ANY}(x, out_dir = NULL)
}
\description{
Can be created with function \code{\link{dmDSdata}}.
}
\section{Slots}{

\describe{
\item{\code{counts}}{\code{\linkS4class{MatrixList}} of counts.}

\item{\code{samples}}{data.frame with information about samples. Contains unique sample names (\code{sample_id}) and information about grouping into conditions (\code{group}).}
}}
\examples{
d <- dataDS_dmDSdata
plot(d)
plotData(d)
}

