% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSdata.R
\name{dmDSdata}
\alias{dmDSdata}
\title{Create dmDSdata object from a table of counts}
\usage{
dmDSdata(counts, gene_id, feature_id, sample_id, group)
}
\arguments{
\item{counts}{A numeric matrix or data.frame of counts. Rows represent features (exons,
  bins or transcripts), columns represent samples.}

\item{gene_id}{Vector of gene IDs of lenght correspoding to the number of rows in \code{counts}.}

\item{feature_id}{Vector of feature IDs of lenght correspoding to the number of rows in \code{counts}.}

\item{sample_id}{A vector of unique sample IDs of length corresponding to
  the number of columns in \code{counts}.}

\item{group}{A vector that defines the goupping of samples.}
}
\value{
Returns \code{dmDSdata} object containing counts and sample
   information \code{\linkS4class{dmDSdata}}.
}
\description{
Create dmDSdata object from a table of counts
}
\examples{
head(dataDS_counts)
 dataDS_metadata

 counts <- dataDS_counts[,-1]
 group_id <- dataDS_counts[,1]
 group_split <- limma::strsplit2(group_id, ":")
 gene_id <- group_split[, 1]
 feature_id <- group_split[, 2]
 sample_id = dataDS_metadata$sample_id
 group = dataDS_metadata$group

 d <- dmDSdata(counts = counts, gene_id = gene_id, feature_id = feature_id, sample_id = sample_id, group = group)
}

