% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dmDSdata.R
\name{dmDSdata}
\alias{dmDSdata}
\title{Create dmDSdata object
 
 Constructor function for a \code{\linkS4class{dmDSdata}} object. 
 
 @param counts Numeric matrix or data frame of counts. Rows represent features, for example, exons,
   exonic bins or transcripts. Columns represent samples.
 @param gene_id Vector of gene IDs corresponding to \code{counts}.
 @param feature_id Vector of feature IDs corresponding to \code{counts}.
 @param sample_id Vector of unique sample IDs corresponding to the columns in \code{counts}.
 @param group Vector that defines the grouping of samples.
 @return Returns a \linkS4class{dmDSdata} object.
 @examples
 
#############################
### Create dmDSdata object
#############################
### Get HTSeq exonic bin counts from 'pasilla' package}
\usage{
dmDSdata(counts, gene_id, feature_id, sample_id, group)
}
\description{
library(pasilla)
}
\details{
data_dir  <- system.file("extdata", package="pasilla")
count_files <- list.files(data_dir, pattern="fb.txt$", full.names=TRUE)
count_files

# Create a data frame with htseq counts
htseq_list <- lapply(1:length(count_files), function(i){
  # i = 1
  htseq <- read.table(count_files[i], header = FALSE, as.is = TRUE)
  colnames(htseq) <- c("group_id", gsub("fb.txt", "", 
     strsplit(count_files[i], "extdata/")[[1]][2]))
  return(htseq)
})

htseq_counts <- Reduce(function(...) merge(..., by = "group_id", all=TRUE, 
   sort = FALSE), htseq_list)
tail(htseq_counts)
htseq_counts <- htseq_counts[!grepl(pattern = "_", htseq_counts$group_id), ]

group_split <- limma::strsplit2(htseq_counts[, 1], ":")

d <- dmDSdata(counts = htseq_counts[, -1], gene_id = group_split[, 1], 
   feature_id = group_split[, 2], sample_id = colnames(htseq_counts)[-1], 
   group = gsub("[1-4]", "", colnames(htseq_counts)[-1]))

plotData(d)

# Use a subset of genes, which is defined in the following file
genes_subset = readLines(file.path(data_dir, "geneIDsinsubset.txt"))
d <- d[names(d) %in% genes_subset, ]

plotData(d)

@seealso \code{\link{plotData}}, \code{\link{dmFilter}}, \code{\link{dmDispersion}}, \code{\link{dmFit}}, \code{\link{dmTest}}
 @author Malgorzata Nowicka
 @export
}

