% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSdata.R
\name{dmDSdata}
\alias{dmDSdata}
\title{Create dmDSdata object from counts}
\usage{
dmDSdata(counts, gene_id_counts, feature_id_counts, sample_id, group)
}
\arguments{
\item{counts}{A numeric matrix of counts. Rows represent features (exons,
bins or transcripts), columns represent samples.}

\item{gene_id_counts}{A vector of gene IDs of lenght correspoding to the
number of rows in \code{counts}.}

\item{feature_id_counts}{A vector of feature IDs of lenght correspoding to
the number of rows in \code{counts}.}

\item{sample_id}{A vector of unique sample IDs of length corresponding to the
number of columns in \code{counts}.}

\item{group}{A vector that defines the goupping of samples.}
}
\value{
Return \code{dmDSdata} object containing counts and sample
  information.
}
\description{
Create dmDSdata object from counts
}

