% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSdata.R
\name{dmDSdata}
\alias{dmDSdata}
\title{Create dmDSdata object}
\usage{
dmDSdata(counts, gene_id, feature_id, sample_id, group)
}
\arguments{
\item{counts}{Numeric matrix or data frame of counts. Rows represent features, for example, exons,
  exonic bins or transcripts. Columns represent samples.}

\item{gene_id}{Vector of gene IDs corresponding to \code{counts}.}

\item{feature_id}{Vector of feature IDs corresponding to \code{counts}.}

\item{sample_id}{Vector of unique sample IDs corresponding to the columns in \code{counts}.}

\item{group}{Vector that defines the goupping of samples.}
}
\value{
Returns a \linkS4class{dmDSdata} object.
}
\description{
Constructor function for a \code{\linkS4class{dmDSdata}} object.
}
\examples{
### Differential splicing analysis
 ### Create dmDSdata object

 # counts
 head(dataDS_counts)
 # metadata
 head(dataDS_metadata)

 group_split <- limma::strsplit2(dataDS_counts[, 1], ":")

 d <- dmDSdata(counts = dataDS_counts[, -1], gene_id = group_split[, 1],
   feature_id = group_split[, 2], sample_id = dataDS_metadata$sample_id,
   group = dataDS_metadata$group)

 plotData(d)
}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{dataDS_counts}}, \code{\link{dataDS_metadata}}, \code{\link{plotData}}, \code{\link{dmFilter}}, \code{\link{dmDispersion}}, \code{\link{dmFit}}, \code{\link{dmTest}}
}

