% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSfit.R
\docType{class}
\name{dmDSfit-class}
\alias{dmDSfit-class}
\alias{plot,dmDSfit,ANY-method}
\alias{proportions}
\alias{proportions,dmDSfit-method}
\alias{statistics}
\alias{statistics,dmDSfit-method}
\title{Object that extends \code{dmDSdispersion} class by adding fitting.}
\usage{
proportions(x, ...)

\S4method{proportions}{dmDSfit}(x)

statistics(x, ...)

\S4method{statistics}{dmDSfit}(x)

\S4method{plot}{dmDSfit,ANY}(x, gene_id, plot_type = "barplot",
  order = TRUE, plot_full = TRUE, plot_main = TRUE, out_dir = NULL)
}
\description{
Object that extends \code{dmDSdispersion} class by adding fitting.
}
\section{Slots}{

\describe{
\item{\code{dispersion}}{Character specifying which type of dispersion was used for fitting: common_dispersion or genewise_dispersion.}

\item{\code{fit_full}}{\code{\linkS4class{MatrixList}} object containing the per group proportions of feature expression and likelihoods.}
}}

