% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSfit.R
\docType{class}
\name{dmDSfit-class}
\alias{dmDSfit-class}
\alias{plot,dmDSfit,ANY-method}
\title{Object that extends \code{dmDSdispersion} class by adding fitting.}
\usage{
\S4method{plot}{dmDSfit,ANY}(x, gene_id, plot_type = "barplot",
  order = TRUE, plot_full = TRUE, out_dir = NULL)
}
\description{
Object that extends \code{dmDSdispersion} class by adding fitting.
}
\section{Slots}{

\describe{
\item{\code{dispersion}}{Character specifying which type of dispersion was used for fitting: common_dispersion or genewise_dispersion.}

\item{\code{fit_full}}{\code{\linkS4class{MatrixList}} object containing the per group proportions of feature expression and likelihoods.}
}}

