% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDStest.R
\docType{class}
\name{dmDStest-class}
\alias{dmDStest-class}
\alias{proportions,dmDStest-method}
\alias{results}
\alias{results,dmDStest-method}
\alias{statistics,dmDStest-method}
\title{Object that extends \code{dmDSfit} by adding the test results.}
\usage{
\S4method{proportions}{dmDStest}(x)

\S4method{statistics}{dmDStest}(x)

results(x, ...)

\S4method{results}{dmDStest}(x)
}
\description{
Object that extends \code{dmDSfit} by adding the test results.
}
\section{Slots}{

\describe{
\item{\code{fit_null}}{list of null models defined by contrasts.}

\item{\code{results}}{data.frame with \code{gene_id} - gene IDs, \code{lr} - likelihood ratio statistics, \code{df} - degrees of freedom, \code{pvalue} - p-values and \code{adj_pvalue} - Benjamini & Hochberg adjusted p-values.}
}}

