% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSfit.R, R/class_dmSQTLfit.R
\docType{methods}
\name{dmFit}
\alias{dmFit}
\alias{dmFit,dmDSdispersion-method}
\alias{dmFit,dmSQTLdispersion-method}
\title{Estimating the proportions and likelihoods of Dirichlet-multinomial.}
\usage{
dmFit(x, ...)

\S4method{dmFit}{dmDSdispersion}(x, dispersion = "genewise_dispersion",
  prop_mode = c("constrOptim", "constrOptimG", "FisherScoring")[2],
  prop_tol = 1e-12, verbose = FALSE,
  BPPARAM = BiocParallel::MulticoreParam(workers = 1))

\S4method{dmFit}{dmSQTLdispersion}(x, dispersion = "genewise_dispersion",
  prop_mode = "constrOptimG", prop_tol = 1e-12, verbose = FALSE,
  BPPARAM = BiocParallel::MulticoreParam(workers = 1))
}
\arguments{
\item{x}{\code{\linkS4class{dmDSdispersion}} or \code{\linkS4class{dmSQTLdispersion}} object or any that inherits from them.}

\item{...}{Parameters needed for the proportion estimation.}

\item{dispersion}{Characted defining which dispersion should be used for fitting. Possible values \code{"genewise_dispersion"}, \code{"common_dispersion"}}

\item{prop_mode}{Optimization method used to estimate the proportions. Possible values \code{"constrOptim", "constrOptimG"}.}

\item{prop_tol}{The desired accuracy when estimating dispersion.}

\item{verbose}{Logical.}

\item{BPPARAM}{Parallelization method used by \code{\link[BiocParallel]{bplapply}}.}
}
\description{
Estimating the proportions and likelihoods of Dirichlet-multinomial.
}
\examples{
d <- dataDS_dmDSdispersion
d <- dmFit(d)
}

