% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSLRT.R, R/class_dmSQTLLRT.R
\docType{methods}
\name{dmLRT}
\alias{dmLRT}
\alias{dmLRT,dmDSfit-method}
\alias{dmLRT,dmSQTLfit-method}
\title{Likelihood ratio test between full and null model.}
\usage{
dmLRT(x, ...)

\S4method{dmLRT}{dmDSfit}(x, compared_groups = 1:nlevels(samples(x)$group),
  prop_mode = "constrOptimG", prop_tol = 1e-12, verbose = FALSE,
  BPPARAM = BiocParallel::MulticoreParam(workers = 1))

\S4method{dmLRT}{dmSQTLfit}(x, prop_mode = "constrOptimG", prop_tol = 1e-12,
  verbose = FALSE, BPPARAM = BiocParallel::MulticoreParam(workers = 1))
}
\arguments{
\item{x}{\code{\linkS4class{dmDSfit}} or \code{\linkS4class{dmSQTLfit}} object.}

\item{...}{Parameters needed for the likelihood ratio test.}

\item{compared_groups}{vector or a list of vectors.}
}
\value{
This function returns a \code{\linkS4class{dmDSLRT}} or \code{\linkS4class{dmSQTLLRT}} object with an additional slot \code{table} which is sorted by significance and contains  \code{gene_id} - gene IDs, \code{lr} - likelihood ratio statistics, \code{df} - degrees of freedom, \code{pvalue} - p-values and \code{adj_pvalue} - Benjamini & Hochberg adjusted p-values.
}
\description{
Likelihood ratio test between full and null model.
}
\examples{
d <- dataDS_dmDSdispersion

# If possible, increase the number of workers
d <- dmFit(d, BPPARAM = BiocParallel::MulticoreParam(workers = 1))

d <- dmLRT(d)

results <- results(d)
}

