% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmSQTLdata.R
\docType{class}
\name{dmSQTLdata-class}
\alias{dmSQTLdata-class}
\title{Object that contains counts, genotypes and sample information.}
\description{
Can be created with function \code{\link{dmSQTLdata}} and \code{\link{dmSQTLdataFromRanges}}.
}
\section{Slots}{

\describe{
\item{\code{counts}}{\code{\linkS4class{MatrixList}}  of counts.}

\item{\code{genotypes}}{\code{\linkS4class{MatrixList}} with genotypes.}

\item{\code{samples}}{data.frame with information about samples. Contains unique sample names (\code{sample_id}).}
}}

