% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmSQTLdata.R
\name{dmSQTLdata}
\alias{dmSQTLdata}
\alias{dmSQTLdataFromRanges}
\title{Create \code{\linkS4class{dmSQTLdata}} object from tables of counts and matched genotypes}
\usage{
dmSQTLdata(counts, gene_id, feature_id, genotypes, gene_id_genotypes, snp_id,
  sample_id, BPPARAM = BiocParallel::MulticoreParam(workers = 1))

dmSQTLdataFromRanges(counts, gene_id, feature_id, gene_ranges, genotypes,
  snp_id, snp_ranges, sample_id, window = 5000,
  BPPARAM = BiocParallel::MulticoreParam(workers = 1))
}
\arguments{
\item{counts}{A numeric matrix or data.frame of counts. Rows represent features (exons,
  bins or transcripts), columns represent samples.}

\item{gene_id}{Vector of gene IDs of lenght correspoding to the number of rows in \code{counts}.}

\item{feature_id}{Vector of feature IDs of lenght correspoding to the number of rows in \code{counts}.}

\item{genotypes}{A numeric matrix of matched genotypes. See Details.}

\item{gene_id_genotypes}{Vector of gene IDs that are matched with genotypes.}

\item{snp_id}{Vector of SNP IDs that correspond to genotypes.}

\item{sample_id}{A vector of unique sample IDs of length corresponding to
  the number of columns in \code{counts}.}

\item{gene_ranges}{\code{\linkS4class{GRanges}} object with information
  about gene location.}

\item{snp_ranges}{\code{\linkS4class{GRanges}} object with information about
  SNP location.}

\item{window}{Numeric. Size of a down and up stream window that is used to
  match SNPs to a gene. See details.}

\item{genotypes}{A numeric matrix of unmatched genotypes. See Details.}
}
\value{
Returns a \code{\linkS4class{dmSQTLdata}} object containing counts, genotypes and sample information.
}
\description{
Create \code{\linkS4class{dmSQTLdata}} object from tables of counts and matched genotypes
}
\examples{
### counts
 head(dataSQTL_counts)

 counts <- dataSQTL_counts[, -(1:2)]
 gene_id <- group_split[, 1]
 feature_id <- group_split[, 2]

 ### gene_ranges
 dataSQTL_gene_ranges
 gene_ranges <- dataSQTL_gene_ranges
 names(gene_ranges) <- S4Vectors::mcols(gene_ranges)$name

 ### genotypes
 head(dataSQTL_genotypes)
 genotypes <- dataSQTL_genotypes[, -(1:4)]

 snp_id <- dataSQTL_genotypes$snp_id

 snp_ranges <- GenomicRanges::GRanges(S4Vectors::Rle(dataSQTL_genotypes$chr), IRanges::IRanges(dataSQTL_genotypes$start, dataSQTL_genotypes$end))
 names(snp_ranges) <- dataSQTL_genotypes$snp_id

 all(colnames(counts) == colnames(genotypes))

 sample_id <- colnames(counts)

 ### create dmSQTLdata object
 d <- dmSQTLdataFromRanges(counts, gene_id, feature_id, gene_ranges, genotypes, snp_id, snp_ranges, sample_id, window = 5e3)
}

