% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmSQTLdispersion.R
\docType{class}
\name{dmSQTLdispersion-class}
\alias{dmSQTLdispersion-class}
\title{Object that extends \code{dmSQTLdata} by adding dipsersion.}
\description{
Object that extends \code{dmSQTLdata} by adding dipsersion.
}
\section{Slots}{

\describe{
\item{\code{mean_expression}}{Numeric vector of mean gene expression.}

\item{\code{common_dispersion}}{Numeric.}

\item{\code{tagwise_dispersion}}{List of estimated tagwise dispersion per gene and genotype.}
}}

