% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmSQTLdispersion.R
\docType{class}
\name{dmSQTLdispersion-class}
\alias{dmSQTLdispersion-class}
\alias{plot,dmSQTLdispersion,ANY-method}
\title{Object that extends \code{dmSQTLdata} by adding dipsersion.}
\usage{
\S4method{plot}{dmSQTLdispersion,ANY}(x, out_dir = NULL)
}
\description{
Object that extends \code{dmSQTLdata} by adding dipsersion.
}
\section{Slots}{

\describe{
\item{\code{mean_expression}}{Numeric vector of mean gene expression.}

\item{\code{common_dispersion}}{Numeric.}

\item{\code{genewise_dispersion}}{List of estimated tagwise dispersion per gene and genotype.}
}}

