% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmSQTLfit.R
\docType{class}
\name{dmSQTLfit-class}
\alias{dmSQTLfit-class}
\title{Object that extends \code{dmSQTLdispersion} by adding fitting.}
\description{
Object that extends \code{dmSQTLdispersion} by adding fitting.
}
\section{Slots}{

\describe{
\item{\code{dispersion}}{Character specifying which type of dispersion was used for fitting.}

\item{\code{fit_full}}{List of \code{\linkS4class{MatrixList}} objects. Each of them contains the per gene fitting results that include the group proportions (groups are defined by genotypes) of feature expression, full model likelihoods and degrees of freedom.}

\item{\code{fit_null}}{List of \code{\linkS4class{MatrixList}} objects. Each of them contsins the per gene fitting results that include the pooled proportions of feature expression, null model likelihoods and degrees of freedom.}
}}

