% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmSQTLtest.R
\docType{class}
\name{dmSQTLtest-class}
\alias{dmSQTLtest-class}
\title{Object that extends \code{dmSQTLfit} by adding the test results.}
\description{
Object that extends \code{dmSQTLfit} by adding the test results.
}
\section{Slots}{

\describe{
\item{\code{table}}{DataFrame with \code{gene_id} - gene IDs, \code{snp_id} - SNP IDs, \code{lr} - likelihood ratio statistics, \code{df} - degrees of freedom, \code{pvalue} - p-values and \code{adj_pvalue} - Benjamini & Hochberg adjusted p-values.}
}}

