% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDStest.R, R/class_dmSQTLtest.R
\docType{methods}
\name{dmTest}
\alias{dmTest}
\alias{dmTest,dmDSfit-method}
\alias{dmTest,dmSQTLfit-method}
\title{Likelihood ratio test}
\usage{
dmTest(x, ...)

\S4method{dmTest}{dmDSfit}(x, compared_groups = 1:nlevels(samples(x)$group),
  prop_mode = "constrOptimG", prop_tol = 1e-12, verbose = FALSE,
  BPPARAM = BiocParallel::MulticoreParam(workers = 1))

\S4method{dmTest}{dmSQTLfit}(x, prop_mode = "constrOptimG",
  prop_tol = 1e-12, verbose = FALSE,
  BPPARAM = BiocParallel::MulticoreParam(workers = 1))
}
\arguments{
\item{x}{\code{\linkS4class{dmDSfit}} or \code{\linkS4class{dmSQTLfit}} object.}

\item{...}{Other parameters that can be defined by methods using this generic.}

\item{compared_groups}{Vector or a list of vectors that defines which experimental conditions should be tested for differential splicing. By default, we test for a difference between any of the groups specified in \code{samples(x)$group}. Values in this vectors should be levels or numbers of the levels of \code{samples(x)$group}.}

\item{prop_mode}{Optimization method used to estimate proportions. Possible values \code{"constrOptim"} and \code{"constrOptimG"}.}

\item{prop_tol}{The desired accuracy when estimating proportions.}

\item{verbose}{Logical. Whether to display more progress messages.}

\item{BPPARAM}{Parallelization method used by \code{\link[BiocParallel]{bplapply}}.}
}
\value{
Returns a \code{\linkS4class{dmDStest}} or \code{\linkS4class{dmSQTLtest}} object.
}
\description{
First, estimate the null Dirichlet-multinomial model proportions, i.e., feature ratios are estimated based on pooled data. Use the likelihood ratio statistic to test for the difference between feature proportions in different groups to identify the differentially spliced genes (differential splicing analysis) or the sQTLs (sQTL analysis).
}
\examples{
### Differential splicing analysis

d <- dataDS_dmDSdispersion

# If possible, increase the number of workers
d <- dmFit(d, BPPARAM = BiocParallel::MulticoreParam(workers = 1))

d <- dmTest(d, BPPARAM = BiocParallel::MulticoreParam(workers = 1))

plotTest(d)

res <- results(d)
res <- res[order(res$pvalue, decreasing = FALSE), ]

plotFit(d, gene_id = res$gene_id[1])
}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{plotTest}}, \code{\link{dmDispersion}}, \code{\link{dmFit}}
}

