% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDStest.R, R/class_dmSQTLtest.R
\docType{methods}
\name{dmTest}
\alias{dmTest}
\alias{dmTest,dmDSfit-method}
\alias{dmTest,dmSQTLfit-method}
\title{Likelihood ratio test}
\usage{
dmTest(x, ...)

\S4method{dmTest}{dmDSfit}(x, compared_groups = levels(samples(x)$group),
  test = "lr", prop_mode = "constrOptimG", prop_tol = 1e-12,
  verbose = 0, BPPARAM = BiocParallel::MulticoreParam(workers = 1))

\S4method{dmTest}{dmSQTLfit}(x, test = "lr", permutations = "all_genes",
  prop_mode = "constrOptimG", prop_tol = 1e-12, verbose = 0,
  BPPARAM = BiocParallel::MulticoreParam(workers = 1))
}
\arguments{
\item{x}{\code{\linkS4class{dmDSfit}} or \code{\linkS4class{dmSQTLfit}} object.}

\item{...}{Other parameters that can be defined by methods using this generic.}

\item{compared_groups}{Vector that defines which experimental conditions should be tested for differential splicing. By default, we test for a difference between any of the groups specified in \code{samples(x)$group}. Values in this vector should indicate levels or numbers of levels in \code{samples(x)$group}.}

\item{test}{Character specifying which test to use. \code{"LR"} - likelihood ratio test, \code{"F"} - F-test.}

\item{prop_mode}{Optimization method used to estimate proportions. Possible values \code{"constrOptim"} and \code{"constrOptimG"}.}

\item{prop_tol}{The desired accuracy when estimating proportions.}

\item{verbose}{Numeric. Definie the level of progress messages displayed. 0 - no messages, 1 - main messages, 2 - message for every gene fitting.}

\item{BPPARAM}{Parallelization method used by \code{\link[BiocParallel]{bplapply}}.}
}
\value{
Returns a \code{\linkS4class{dmDStest}} or \code{\linkS4class{dmSQTLtest}} object.
}
\description{
First, estimate the null Dirichlet-multinomial model proportions, i.e., feature ratios are estimated based on pooled (no grouping into conditions) counts. Use the likelihood ratio statistic to test for the difference between feature proportions in different groups to identify the differentially spliced genes (differential splicing analysis) or the sQTLs (sQTL analysis).
}
\examples{
###################################
### Differential splicing analysis
###################################
# If possible, increase the number of workers in BPPARAM

d <- data_dmDSdata
\donttest{
### Filtering
# Check what is the minimal number of replicates per condition
table(samples(d)$group)
d <- dmFilter(d, min_samps_gene_expr = 6, min_samps_feature_expr = 3,
 min_samps_feature_prop = 3)

### Calculate dispersion
d <- dmDispersion(d, BPPARAM = BiocParallel::MulticoreParam(workers = 1))

### Fit full model proportions
d <- dmFit(d, BPPARAM = BiocParallel::MulticoreParam(workers = 1))

### Fit null model proportions and test for DS
d <- dmTest(d, BPPARAM = BiocParallel::MulticoreParam(workers = 1))
plotTest(d)

head(proportions(d))
head(statistics(d))
head(results(d))

}
}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{data_dmDSdata}}, \code{\link{data_dmSQTLdata}}, \code{\link{plotTest}}, \code{\link{dmDispersion}}, \code{\link{dmFit}}
}

