% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_MatrixList.R
\docType{class}
\name{MatrixList-class}
\alias{$,MatrixList-method}
\alias{MatrixList-class}
\alias{[,MatrixList,ANY,ANY-method}
\alias{[[,MatrixList,ANY,ANY-method}
\alias{colnames,MatrixList-method}
\alias{dim,MatrixList-method}
\alias{length,MatrixList-method}
\alias{names,MatrixList-method}
\alias{ncol,MatrixList-method}
\alias{nrow,MatrixList-method}
\alias{rownames,MatrixList-method}
\alias{width,MatrixList-method}
\title{MatrixList container}
\usage{
\S4method{names}{MatrixList}(x)

\S4method{rownames}{MatrixList}(x)

\S4method{colnames}{MatrixList}(x)

\S4method{length}{MatrixList}(x)

\S4method{width}{MatrixList}(x)

\S4method{dim}{MatrixList}(x)

\S4method{nrow}{MatrixList}(x)

\S4method{ncol}{MatrixList}(x)

\S4method{[[}{MatrixList,ANY,ANY}(x, i)

\S4method{$}{MatrixList}(x, name)

\S4method{[}{MatrixList,ANY,ANY}(x, i, j)
}
\description{
MatrixList container
}
\section{Slots}{

\describe{
\item{\code{unlistData}}{Matrix data}

\item{\code{partitioning}}{List with indexes of each submatrix}

\item{\code{metadata}}{Matrix of additional information}
}}

