% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_MatrixList.R
\docType{class}
\name{MatrixList-class}
\alias{$,MatrixList-method}
\alias{MatrixList-class}
\alias{[,MatrixList,ANY,ANY-method}
\alias{[[,MatrixList,ANY,ANY-method}
\alias{colnames,MatrixList-method}
\alias{colnames<-,MatrixList-method}
\alias{dim,MatrixList-method}
\alias{length,MatrixList-method}
\alias{names,MatrixList-method}
\alias{names<-,MatrixList-method}
\alias{ncol,MatrixList-method}
\alias{nrow,MatrixList-method}
\alias{rownames,MatrixList-method}
\alias{rownames<-,MatrixList-method}
\alias{width,MatrixList-method}
\title{MatrixList object}
\usage{
\S4method{names}{MatrixList}(x)

\S4method{names}{MatrixList}(x) <- value

\S4method{rownames}{MatrixList}(x)

\S4method{rownames}{MatrixList}(x) <- value

\S4method{colnames}{MatrixList}(x)

\S4method{colnames}{MatrixList}(x) <- value

\S4method{length}{MatrixList}(x)

\S4method{width}{MatrixList}(x)

\S4method{dim}{MatrixList}(x)

\S4method{nrow}{MatrixList}(x)

\S4method{ncol}{MatrixList}(x)

\S4method{[[}{MatrixList,ANY,ANY}(x, i)

\S4method{$}{MatrixList}(x, name)

\S4method{[}{MatrixList,ANY,ANY}(x, i, j)
}
\arguments{
\item{x}{MatrixList object.}

\item{value}{Values used for the replacement of corresponding attributes.}

\item{i,j}{Indexes of MatrixList to be subsampled.}

\item{name}{Name of submatrix to be accessed.}
}
\description{
A MatrixList object is a 'list' container for matrices which have the same number of columns but varying number of rows. Columns of these matrices may or may not have the same interpretation. Additionally, one can store an extra information corresponding to each of the matrices in 'metadata' matrix.
}
\section{Slots}{

\describe{
\item{\code{unlistData}}{A matrix which is a row binding of all the stored matrices.}

\item{\code{partitioning}}{A list of indexes which defines the row partitioning of 'unlistData' matrix into the original submatrices.}

\item{\code{metadata}}{A matrix of additional information where each row corresponds to one of the submatrices.}
}}

