% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSdata.R, R/class_dmSQTLdata.R
\docType{methods}
\name{plotData}
\alias{plotData}
\alias{plotData,dmDSdata-method}
\alias{plotData,dmSQTLdata-method}
\title{Plot data summary}
\usage{
plotData(x, ...)

\S4method{plotData}{dmDSdata}(x, out_dir = NULL)

\S4method{plotData}{dmSQTLdata}(x, out_dir = NULL)
}
\arguments{
\item{x}{\code{\linkS4class{dmDSdata}} or \code{\linkS4class{dmSQTLdata}} object.}

\item{...}{Other parameters that can be defined by methods using this generic.}

\item{out_dir}{Directory where the plot should be saved. If \code{NULL} the plot is printed.}
}
\description{
Plot a histogram of the number of features per gene. Additionally, for \code{\linkS4class{dmSQTLdata}} object, plot a histogram of the number of SNPs per gene and a histogram of the number of unique SNPs (blocks) per gene.
}
\examples{
### dmDSdata

d <- dataDS_dmDSdata
plotData(d)
plot(d)
### dmSQTLdata

d <- dataSQTL_dmSQTLdata
plotData(d)
plot(d)
}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{plotDispersion}}, \code{\link{plotFit}}, \code{\link{plotLRT}}
}

