% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSdispersion.R, R/class_dmSQTLdispersion.R
\docType{methods}
\name{plotDispersion}
\alias{plotDispersion}
\alias{plotDispersion,dmDSdispersion-method}
\alias{plotDispersion,dmSQTLdispersion-method}
\title{Plot the dispersion - mean trend.}
\usage{
plotDispersion(x, ...)

\S4method{plotDispersion}{dmDSdispersion}(x, out_dir = NULL)

\S4method{plotDispersion}{dmSQTLdispersion}(x, out_dir = NULL)
}
\arguments{
\item{x}{\code{\linkS4class{dmDSdispersion}} or \code{\linkS4class{dmSQTLdispersion}} object or any that inherits from it.}

\item{...}{Plotting parameters.}

\item{out_dir}{Directory where the plot should be saved. If \code{NULL} the plot is printed.}
}
\description{
Plot the dispersion - mean trend.
}
\examples{
d <- dataDS_dmDSdispersion
plotDispersion(d)
plot(d)
}

