% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSfit.R, R/class_dmDStest.R, R/class_dmSQTLfit.R, R/class_dmSQTLtest.R
\docType{methods}
\name{plotFit}
\alias{plotFit}
\alias{plotFit,dmDSfit-method}
\alias{plotFit,dmDStest-method}
\alias{plotFit,dmSQTLfit-method}
\alias{plotFit,dmSQTLtest-method}
\title{Plot feature proportions}
\usage{
plotFit(x, ...)

\S4method{plotFit}{dmDSfit}(x, gene_id, plot_type = "barplot", order = TRUE,
  plot_full = TRUE, plot_main = TRUE, out_dir = NULL)

\S4method{plotFit}{dmDStest}(x, gene_id, plot_type = "barplot",
  order = TRUE, plot_full = TRUE, plot_null = TRUE, compared_groups = 1,
  plot_main = TRUE, out_dir = NULL)

\S4method{plotFit}{dmSQTLfit}(x, gene_id, snp_id, plot_type = "boxplot1",
  order = TRUE, plot_full = TRUE, plot_main = TRUE, out_dir = NULL)

\S4method{plotFit}{dmSQTLtest}(x, gene_id, snp_id, plot_type = "boxplot1",
  order = TRUE, plot_full = TRUE, plot_null = TRUE, plot_main = TRUE,
  out_dir = NULL)
}
\arguments{
\item{x}{\code{\linkS4class{dmDSfit}}, \code{\linkS4class{dmDStest}} or \code{\linkS4class{dmSQTLfit}}, \code{\linkS4class{dmSQTLtest}} object.}

\item{...}{Other parameters that can be defined by methods using this generic.}

\item{gene_id}{Vector of gene IDs to be plotted.}

\item{plot_type}{Character defining the type of the plot produced. Possible values \code{"barplot"}, \code{"boxplot1"}, \code{"boxplot2"}, \code{"lineplot"}, \code{"ribbonplot"}.}

\item{order}{Logical. Whether to plot the features ordered by their expression.}

\item{plot_full}{Logical. Whether to plot the proportions estimated by the full model.}

\item{plot_main}{Logical. Whether to plot a title with the information about the Dirichlet-multinomial estimates.}

\item{out_dir}{Directory where the plot should be saved. If \code{NULL} the plot is printed.}

\item{plot_null}{Logical. Whether to plot the proportions estimated by the null model.}

\item{compared_groups}{Numeric or character indicating comparison, defined in \code{\link{dmTest}}, that should be plotted.}

\item{snp_id}{Vector of SNP IDs to be plotted. \code{snp_id} must match \code{gene_id}.}
}
\value{
Plot, per gene, the observed and estimated with Dirichlet-multinomial model feature ratios. Estimated proportions are marked with diamond shapes.
}
\description{
Plot feature proportions
}
\examples{
### Differential splicing analysis
### Plot feature proportions for top DS gene

d <- dataDS_dmDStest

res <- results(d)
res <- res[order(res$pvalue, decreasing = FALSE), ]

gene_id <- res$gene_id[1]

plotFit(d, gene_id = gene_id)
plotFit(d, gene_id = gene_id, plot_type = "lineplot")
plotFit(d, gene_id = gene_id, plot_type = "ribbonplot")
}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{plotData}}, \code{\link{plotDispersion}}, \code{\link{plotTest}}
}

