% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSfit.R, R/class_dmDStest.R, R/class_dmSQTLfit.R, R/class_dmSQTLtest.R
\docType{methods}
\name{plotFit}
\alias{plotFit}
\alias{plotFit,dmDSfit-method}
\alias{plotFit,dmDStest-method}
\alias{plotFit,dmSQTLfit-method}
\alias{plotFit,dmSQTLtest-method}
\title{Plot feature proportions}
\usage{
plotFit(x, ...)

\S4method{plotFit}{dmDSfit}(x, gene_id, plot_type = "barplot", order = TRUE,
  plot_full = TRUE, plot_main = TRUE, out_dir = NULL)

\S4method{plotFit}{dmDStest}(x, gene_id, plot_type = "barplot",
  order = TRUE, plot_full = TRUE, plot_null = TRUE, compared_groups = 1,
  plot_main = TRUE, out_dir = NULL)

\S4method{plotFit}{dmSQTLfit}(x, gene_id, snp_id, plot_type = "boxplot1",
  order = TRUE, plot_full = TRUE, plot_main = TRUE, out_dir = NULL)

\S4method{plotFit}{dmSQTLtest}(x, gene_id, snp_id, plot_type = "boxplot1",
  order = TRUE, plot_full = TRUE, plot_null = TRUE, plot_main = TRUE,
  out_dir = NULL)
}
\arguments{
\item{x}{\code{\linkS4class{dmDSfit}} or \code{\linkS4class{dmSQTLfit}} object.}

\item{...}{Other parameters that can be defined by methods using this generic.}

\item{gene_id}{Vector of gene IDs to be plotted.}

\item{plot_type}{Character defining the type of the plot produced. Possible values \code{"barplot"}, \code{"boxplot1"}, \code{"boxplot2"}, \code{"lineplot"}, \code{"ribbonplot"}.}

\item{order}{Logical. Whether to plot the features ordered by their expression.}

\item{plot_full}{Logical. Whether to plot the proportions estimated by the full model.}

\item{plot_main}{Logical. Whether to plot a title with the information about the estimates.}

\item{out_dir}{Directory where the plot should be saved. If \code{NULL} the plot is printed.}

\item{compared_groups}{Numeric or character indicating comparison that should be plotted.}
}
\description{
Plot the observed and estimated feature ratios.
}
\examples{
d <- dataDS_dmDSdispersion

# If possible, increase the number of workers
d <- dmFit(d, BPPARAM = BiocParallel::MulticoreParam(workers = 1))

gene_id <- names(d)[1]

plotFit(d, gene_id = gene_id)


### Plot proportions of top 3 genes
d <- dmLRT(d)

res <- results(d)
res <- res[order(res$pvalue, decreasing = FALSE), ]

gene_id <- res$gene_id[1:3]

plotFit(d, gene_id = gene_id, plot_type = "barplot")
plotFit(d, gene_id = gene_id, plot_type = "lineplot")
plotFit(d, gene_id = gene_id, plot_type = "ribbonplot")
}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{plotData}}, \code{\link{plotDispersion}}, \code{\link{plotLRT}}
}

