% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDStest.R, R/class_dmSQTLtest.R
\docType{methods}
\name{plotTest}
\alias{plotTest}
\alias{plotTest,dmDStest-method}
\alias{plotTest,dmSQTLtest-method}
\title{Plot the histogram of p-values.}
\usage{
plotTest(x, ...)

\S4method{plotTest}{dmDStest}(x, out_dir = NULL)

\S4method{plotTest}{dmSQTLtest}(x, out_dir = NULL)
}
\arguments{
\item{x}{\code{\linkS4class{dmDStest}} or \code{\linkS4class{dmSQTLtest}} object.}

\item{...}{Other parameters that can be defined by methods using this generic.}

\item{out_dir}{Directory where the plot should be saved. If \code{NULL} the plot is printed.}
}
\description{
Plot the histogram of p-values.
}

