% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dmDStest.R, R/class_dmSQTLtest.R
\docType{methods}
\name{plotTest}
\alias{plotTest}
\alias{plotTest,dmDStest-method}
\alias{plotTest,dmSQTLtest-method}
\title{Plot p-values distribution}
\usage{
plotTest(x, ...)

\S4method{plotTest}{dmDStest}(x, out_dir = NULL)

\S4method{plotTest}{dmSQTLtest}(x, out_dir = NULL)
}
\arguments{
\item{x}{\code{\linkS4class{dmDStest}} or \code{\linkS4class{dmSQTLtest}} object.}

\item{...}{Other parameters that can be defined by methods using this generic.}

\item{out_dir}{Directory where the plot should be saved. If \code{NULL} the plot is printed.}
}
\value{
Plot a histogram of p-values.
}
\description{
Plot p-values distribution
}
\examples{

###################################
### Differential splicing analysis
###################################
# If possible, increase the number of workers in BPPARAM

d <- data_dmDSdata
\donttest{
### Filtering
# Check what is the minimal number of replicates per condition 
table(samples(d)$group)
d <- dmFilter(d, min_samps_gene_expr = 6, min_samps_feature_expr = 3, 
 min_samps_feature_prop = 3)

### Calculate dispersion
d <- dmDispersion(d, BPPARAM = BiocParallel::MulticoreParam(workers = 1))

### Fit full model proportions
d <- dmFit(d, BPPARAM = BiocParallel::MulticoreParam(workers = 1))

### Fit null model proportions and test for DS
d <- dmTest(d, BPPARAM = BiocParallel::MulticoreParam(workers = 1))

plotTest(d)


}
}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{data_dmDSdata}}, \code{\link{data_dmSQTLdata}}, \code{\link{plotData}}, \code{\link{plotDispersion}}, \code{\link{plotFit}}
}

